+++ app/components/location-sidebar.hbs
@@ -1,0 +1,13 @@
+<Sidebar>
+  <div>
+    <LinkTo @route="locations.new">
+      <Button type="button">
+        {{t "location.new"}}
+      </Button>
+    </LinkTo>
+    <Button {{on "click" this.seed}}> Z A A D </Button>
+  </div>
+  {{#each @model.locations as |location|}}
+    <Sidebar::Location @location={{location}} @didLocationChange={{this.refresh}}/>
+  {{/each}}
+</Sidebar>
