+++ app/routes/initiatives/initiative/edit.js
@@ -1,0 +1,4 @@
+import Route from '@ember/routing/route';
+
+export default class InitiativesInitiativeEditRoute extends Route {
+}
