+++ app/routes/initiatives/new.js
@@ -1,0 +1,4 @@
+import Route from '@ember/routing/route';
+
+export default class InitiativesNewRoute extends Route {
+}
