+++ app/templates/initiatives.hbs
@@ -1,1 +1,4 @@
-{{outlet}}
+<LocationSidebar @model={{@model}}/>
+<div class="w-75 fr">
+  {{outlet}}
+</div>
