+++ app/templates/initiatives/initiative.hbs
@@ -1,1 +1,3 @@
-{{outlet}}
+<div class="pa3 overflow-auto absolute-l bottom-0-l top-0-l right-0-l w-75">
+  {{outlet}}
+</div>
