+++ app/templates/initiatives/initiative/edit.hbs
@@ -1,0 +1,1 @@
+{{outlet}}
