+++ app/templates/initiatives/new.hbs
@@ -1,0 +1,1 @@
+{{outlet}}
