+++ tests/unit/routes/initiatives/initiative/edit-test.js
@@ -1,0 +1,11 @@
+import { module, test } from 'qunit';
+import { setupTest } from 'ember-qunit';
+
+module('Unit | Route | initiatives/initiative/edit', function(hooks) {
+  setupTest(hooks);
+
+  test('it exists', function(assert) {
+    let route = this.owner.lookup('route:initiatives/initiative/edit');
+    assert.ok(route);
+  });
+});
